ffi: Ignore vararg tests with float param
MRI ruby ffi treads floats (only) for varargs as double. But Topaz float
is in all cases an alias for float32. The ffi specs depend on the exception
for floats on vargs.

As I have no idee why is exception was made and now real clue where it would
by needed in topaz, exclude as tests with varargs + float for now.